+++ sdkjs-plugins/v1/onlyoffice-types/src/config/index.d.ts
@@ -1,0 +1,10 @@
+export {
+    type EditorType,
+    type PluginConfig,
+    type VariationConfig,
+    type VariationType,
+    type ButtonConfig,
+    type StoreConfig,
+    type IconConfig,
+    type IconScale,
+} from "./plugin-config";