+++ sdkjs-plugins/v1/onlyoffice-types/src/config/plugin-config.d.ts
@@ -1,0 +1,10 @@
+export type {
+    EditorType,
+    PluginConfig,
+    VariationConfig,
+    VariationType,
+    ButtonConfig,
+    StoreConfig,
+    IconConfig,
+    IconScale,
+} from "../../index";