+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/events.d.ts
@@ -1,0 +1,9 @@
+export type {
+    PluginEventMap,
+    PluginEventHandler,
+    ContextMenuShowEvent,
+    PluginEventName,
+    PluginEventCallback,
+    PluginEditorEventName,
+    PluginEditorEventCallback,
+} from "../../index";