+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/windows.d.ts
@@ -1,0 +1,1 @@
+export type { PluginWindow } from "../../index";