+++ sdkjs-plugins/v1/onlyoffice-types/src/services/desktop-editor.d.ts
@@ -1,0 +1,1 @@
+export type { AscDesktopEditor } from "../../index";