+++ sdkjs-plugins/v1/onlyoffice-types/src/services/simple-request.d.ts
@@ -1,0 +1,4 @@
+export type {
+    AscSimpleRequest,
+    AscSimpleRequestOptions,
+} from "../../index";